+++ crates/dirk_renderer/src/shaders.rs
@@ -1,28 +1,25 @@
-//! This crate contains all the shaders used in the engine.
-
-pub mod types;
+//! Compiled shader blobs used by the renderer.
 
 use std::ffi::CStr;
 
-/// A simple struct that holds a block of shader bytecode and
-/// the name of the shader's entrypoint.
+/// A block of shader bytecode and the shader entry point name.
 pub struct Shader {
     code: &'static [u8],
     entrypoint: &'static CStr,
 }
 
 impl Shader {
-    /// Returns the shader code
+    /// Returns the shader code.
     #[must_use]
     pub const fn code(&self) -> &[u8] {
         self.code
     }
-    /// Returns the code but in blocks of u32
+
+    /// Returns the shader code as little-endian `u32` words.
     ///
     /// # Panics
     ///
-    /// If the SPIR-V code size is not a multiple of
-    /// 4. This would be invalid SPIR-V
+    /// Panics if the SPIR-V code size is not a multiple of 4 bytes.
     #[must_use]
     pub fn code_as_u32(&self) -> Vec<u32> {
         assert!(
@@ -31,10 +28,11 @@ impl Shader {
         );
         self.code
             .chunks_exact(4)
-            .map(|c| u32::from_le_bytes(c.try_into().expect("4 byte chunks")))
+            .map(|chunk| u32::from_le_bytes(chunk.try_into().expect("4 byte chunks")))
             .collect()
     }
-    /// Returns the entrypoint of this shader
+
+    /// Returns the shader entry point.
     #[must_use]
     pub const fn entrypoint(&self) -> &CStr {
         self.entrypoint
@@ -50,5 +48,8 @@ macro_rules! shader {
     };
 }
 
-mod blobs;
-pub use blobs::*;
+/// Vertex shader.
+pub const VERT: Shader = shader!("main_vs", c"main_vs");
+
+/// Fragment shader.
+pub const FRAG: Shader = shader!("main_fs", c"main_fs");
