+++ crates/dirk_shaders/src/lib.rs
@@ -1,0 +1,5 @@
+//! Shared types used by the engine's shaders.
+
+#![no_std]
+
+pub mod types;
